+++ src/parser/inline/media.ts
@@ -9,14 +9,14 @@ import DOM from 'typed-dom';
 
 type SubParsers = [TextParser];
 
-const syntax = /^!\[[^\n]*?\]\(/;
+const syntax = /^!\[[^\n]*?\]\n?\(/;
 
 export const media: MediaParser = function (source: string): Result<HTMLImageElement, never> {
   if (!source.startsWith('![') || source.search(syntax) !== 0) return;
-  const [[, , ...first], next] = loop(combine<SubParsers, HTMLElement | Text>([text]), /^\]\(|^\n/)(source) || [[], ''];
-  if (!next.startsWith('](')) return;
+  const [[, , ...first], next] = loop(combine<SubParsers, HTMLElement | Text>([text]), /^\]\n?\(|^\n/)(source) || [[], ''];
+  if (!next.startsWith('](') && !next.startsWith(']\n(')) return;
   const caption = first.reduce((s, c) => s + c.textContent, '').trim();
-  const [[, , ...second], rest] = loop(text, /^\)|^\s/)(next) || [[], ''];
+  const [[...second], rest] = loop(text, /^\)|^\s/)(next.replace(/^\]\n?\(/, '')) || [[], ''];
   if (!rest.startsWith(')')) return;
   const url = sanitize(second.reduce((s, c) => s + c.textContent, ''));
   if (url === '') return;