+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.23.0
 
 - Extend link syntax.
+- Extend media syntax.
 - Fix media url parsing.
 
 ## 0.22.0